feat: test integration updated
