--- tests/EndToEnd/admin/AGET.test.ts
@@ -1,1 +1,0 @@
-///
