--- tests/EndToEnd/admin/APUT.test.ts
@@ -1,1 +1,0 @@
-///
