--- tests/EndToEnd/driver/DDELETE.test.ts
@@ -1,1 +1,0 @@
-/// /
