--- tests/EndToEnd/packages/PPOST.test.ts
@@ -1,1 +1,0 @@
-/// /
