+++ tests/Integration/driver/DIntEdit.test.ts
@@ -1,0 +1,15 @@
+describe('Edit Driver', () => {
+    const mockDriverData = {
+        username: 'new driver',
+        email: 'newdriver@gmail.com',
+        password: 'newDriver123',
+        phone_number: '1123543223',
+        status: true,
+        profile_pic: '',
+        packages: [],
+    }
+
+    it('', async () => {
+        expect(mockDriverData).toBeDefined()
+    })
+})
