+++ tests/Integration/order/OIntDelete.test.ts
@@ -1,1 +1,10 @@
-///
+describe('Delete Order', () => {
+    const mockOrder = {
+        status: 'Pendiente',
+        driver: 'iug3iu384893gijb',
+        packages: ['obd3oub90b84n4ifhn', 'mhf04b98b3ibf43ofb'],
+    }
+    it('', async () => {
+        expect(mockOrder).toBeDefined()
+    })
+})
